src/lite_modeler/requirements.txt
numpy==1.26.4
Pillow==10.3.0
tflite_runtime==2.14.0
